Move gyro to sensors folder and delete hardware folder
