+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Sensors/OrbitGyro.java
@@ -1,5 +1,5 @@
 
-package org.firstinspires.ftc.teamcode.hardware;
+package org.firstinspires.ftc.teamcode.Sensors;
 
 import com.qualcomm.hardware.bosch.BNO055IMU;
 import com.qualcomm.hardware.bosch.JustLoggingAccelerationIntegrator;
