+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robotData/Constants.java
@@ -3,9 +3,9 @@ package org.firstinspires.ftc.teamcode.robotData;
 public  class Constants {
     public static final float INF = 1e5f;
     public static final float epsilon = 1e-5f;
-    public static final float radius = 0.17288f;
-    public static final float length = 0.3465f;
-    public static final float width = 0.399f;
+    public static final float robotRadius = 0.17288f;
+    public static final float robotLength = 0.3465f;
+    public static final float robotWidth = 0.399f;
     public static final float tileLength = 0.6096f;
 
 }
